requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
cosl==0.0.8
    # via -r requirements.in
hvac==2.1.0
    # via -r requirements.in
idna==3.6
    # via requests
jinja2==3.1.3
    # via -r requirements.in
markupsafe==2.1.4
    # via jinja2
ops==2.10.0
    # via
    #   -r requirements.in
    #   cosl
pydantic==1.10.14
    # via -r requirements.in
pyhcl==0.4.5
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   cosl
    #   ops
requests==2.31.0
    # via hvac
typing-extensions==4.9.0
    # via
    #   cosl
    #   pydantic
urllib3==2.2.0
    # via requests
websocket-client==1.7.0
    # via ops
